Bug 1898506 - Measure the entropy of generated executable code. r=
JIT Spraying mitigations goals are to make all the code which is mapped as
executable to have the least amount of predictibility. The goal being to make it
implausible that a malicious actor can pull-up an exploit abusing the various
Just In Time compilers.

This code introduces a function at the latest stage of all JIT, when the code is
copied to the executable memory, in order to verify whether the generated code
has enough entropy.

Differential Revision: https://phabricator.services.mozilla.com/D211545