Bug 1886559 part 3 - Add the Executable structure. r=jandem
Previously the ExecutableAllocator would return a code pointer as well as using
an out-param to return the ExecutablePool in which the code section got
reserved.

This patch changes this logic to prevent us from splitting the code pointer from
the ExecutablePool by creating an Exectuable structure which holds both
pointers, as well as using move-constructors to ensure the uniqueness of each
allocation.