Bug 1886559 part 4 - Add ExecutableDesc structure. r=jandem
To describe an allocation one need the size and the kind of allocation. Moreover
these informations are also required when finalizing a JitCode, to reduce the
accounting of the code which is no longer used.

Making this ExecutableDesc as part of the Executable class, we ensure that any
allocated pages would always carry the information necessary to finalize it.