Bug 1886559 - JIT: Move constant table out of the code section. r=jandem
Previously, floating point constants were generated at the end of the generated
code. This made it trivial for attackers to make use of our code generator for
injecting any code, before finding a way to locate its position and a way to
jump to it.

This change, as well as the previous one are aimed at making this task more
difficult by moving all constants into memory pages which are never mapped as
executable.

As of this patch this feature only works with JavaScript JITs, which are making
use of the Linker, but not with any of our WASM generated code which are
directly poking at the ExecutableAllocator.