Bug 1886557 - Write JitCode pointers in the data section instead of reading code pointer from the code.
To avoid reading code pointers from the generated code, we bake-in JitCode
pointer in the data section, almost like an ordinary gc::Cell pointer, except
that we are not taking code offset to patch the code once it is linked.

Also, this code should remove the need for the JitCode::FromExecutable function
which is needed by TraceJumpRelocations, which should be made useless by this change.