Bug 1886557 - WIP: Disable X-Only for reading invalidation metadata.
When invalidating Ion code, we patch the code to have some form of weak-pointer
to the IonScript which are holding the memory of the Jit-ed scripts alives,
while they are referenced by the stack. This works well when the executable code
is readable, but fails when we move to X-Only.

The work-around would be to use a mechanism similar to what has been done for
Fuse, using the DependentScriptSet as a table of weak references, as long as any
of the return addresses are on the stack.

While this is doable, this would take too much time to consider as part of this
prototype and to ensure that no error get introduced in the process...